Cart component added for checkoutPage for dry
